DB処理の部分をファイル分け
